Removed ProjectSampleSpecification and moved added code over to ProjectSampleJoinSpecification, and updated existing calls in SampleServiceImpl to build Specification directly.
