+++ src/main/java/ca/corefacility/bioinformatics/irida/repositories/specification/ProjectSampleJoinSpecification.java
@@ -1,39 +1,102 @@
 package ca.corefacility.bioinformatics.irida.repositories.specification;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
+import javax.persistence.criteria.*;
 
 import org.springframework.data.jpa.domain.Specification;
 
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectSampleJoin;
-import ca.corefacility.bioinformatics.irida.model.project.Project;
 
 /**
  * Specification for searching {@link ProjectSampleJoin}s
- * 
- *
  */
-public class ProjectSampleJoinSpecification {
+public class ProjectSampleJoinSpecification implements Specification<ProjectSampleJoin> {
+
+	private List<SearchCriteria> list;
+
+	public ProjectSampleJoinSpecification() {
+		this.list = new ArrayList<>();
+	}
 
 	/**
-	 * Get a {@link ProjectSampleJoin} where the sample name contains the given
-	 * search string and exists in a given Project
-	 * 
-	 * @param name
-	 *            The name to search
-	 * @param project
-	 *            The project to search
-	 * @return a {@link Specification} to use in the repository.
+	 * Add a {@link SearchCriteria}
+	 *
+	 * @param criteria the {@link SearchCriteria} to add to the Specification.
 	 */
-	public static Specification<ProjectSampleJoin> searchSampleWithNameInProject(String name, Project project) {
-		return new Specification<ProjectSampleJoin>() {
-			@Override
-			public Predicate toPredicate(Root<ProjectSampleJoin> root, CriteriaQuery<?> query, CriteriaBuilder cb) {
-				return cb.and(cb.like(root.get("sample").get("sampleName"), "%" + name + "%"),
-						cb.equal(root.get("project"), project));
+	public void add(SearchCriteria criteria) {
+		list.add(criteria);
+	}
+
+	@Override
+	public Predicate toPredicate(Root<ProjectSampleJoin> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
+		//create a new predicate list
+		List<Predicate> predicates = new ArrayList<>();
+
+		//add criteria to predicates
+		for (SearchCriteria criteria : list) {
+			Path<?> path = getPath(root, criteria.getKey());
+
+			if (criteria.getOperation().equals(SearchOperation.GREATER_THAN)) {
+				if (path.getJavaType() == Date.class) {
+					predicates.add(builder.greaterThan(path.as(Date.class), (Date) criteria.getValue()));
+				} else {
+					predicates.add(builder.greaterThan(path.as(String.class), criteria.getValue().toString()));
+				}
+			} else if (criteria.getOperation().equals(SearchOperation.LESS_THAN)) {
+				if (path.getJavaType() == Date.class) {
+					predicates.add(builder.lessThan(path.as(Date.class), (Date) criteria.getValue()));
+				} else {
+					predicates.add(builder.lessThan(path.as(String.class), criteria.getValue().toString()));
+				}
+			} else if (criteria.getOperation().equals(SearchOperation.GREATER_THAN_EQUAL)) {
+				if (path.getJavaType() == Date.class) {
+					predicates.add(builder.greaterThanOrEqualTo(path.as(Date.class), (Date) criteria.getValue()));
+				} else {
+					predicates.add(builder.greaterThanOrEqualTo(path.as(String.class), criteria.getValue().toString()));
+				}
+			} else if (criteria.getOperation().equals(SearchOperation.LESS_THAN_EQUAL)) {
+				if (path.getJavaType() == Date.class) {
+					predicates.add(builder.lessThanOrEqualTo(path.as(Date.class), (Date) criteria.getValue()));
+				} else {
+					predicates.add(builder.lessThanOrEqualTo(path.as(String.class), criteria.getValue().toString()));
+				}
+			} else if (criteria.getOperation().equals(SearchOperation.NOT_EQUAL)) {
+				predicates.add(builder.notEqual(path, criteria.getValue()));
+			} else if (criteria.getOperation().equals(SearchOperation.EQUAL)) {
+				predicates.add(builder.equal(path, criteria.getValue()));
+			} else if (criteria.getOperation().equals(SearchOperation.MATCH)) {
+				predicates.add(builder.like(builder.lower(path.as(String.class)),
+						"%" + criteria.getValue().toString().toLowerCase() + "%"));
+			} else if (criteria.getOperation().equals(SearchOperation.MATCH_END)) {
+				predicates.add(builder.like(builder.lower(path.as(String.class)),
+						criteria.getValue().toString().toLowerCase() + "%"));
+			} else if (criteria.getOperation().equals(SearchOperation.MATCH_START)) {
+				predicates.add(builder.like(builder.lower(path.as(String.class)),
+						"%" + criteria.getValue().toString().toLowerCase()));
+			} else if (criteria.getOperation().equals(SearchOperation.IN)) {
+				predicates.add(builder.in((Path) path).value(criteria.getValue()));
+			} else if (criteria.getOperation().equals(SearchOperation.NOT_IN)) {
+				predicates.add(builder.not((Path) path).in(criteria.getValue()));
 			}
-		};
+		}
+
+		return builder.and(predicates.toArray(new Predicate[0]));
+	}
+
+	private Path<?> getPath(Root<ProjectSampleJoin> root, String attributeName) {
+		Path<?> path;
+		if (attributeName.contains(".")) {
+			String[] split = attributeName.split("\\.");
+			path = root.get(split[0]);
+			for (int i = 1; i < split.length; i++) {
+				path = path.get(split[i]);
+			}
+		} else {
+			path = root.get(attributeName);
+		}
+		return path;
 	}
 }
