+++ src/main/java/ca/corefacility/bioinformatics/irida/service/impl/sample/SampleServiceImpl.java
@@ -56,11 +56,14 @@ import ca.corefacility.bioinformatics.irida.repositories.sample.QCEntryRepositor
 import ca.corefacility.bioinformatics.irida.repositories.sample.SampleRepository;
 import ca.corefacility.bioinformatics.irida.repositories.sequencefile.SequencingObjectRepository;
 import ca.corefacility.bioinformatics.irida.repositories.specification.ProjectSampleJoinSpecification;
-import ca.corefacility.bioinformatics.irida.repositories.specification.ProjectSampleSpecification;
+import ca.corefacility.bioinformatics.irida.repositories.specification.SearchCriteria;
+import ca.corefacility.bioinformatics.irida.repositories.specification.SearchOperation;
 import ca.corefacility.bioinformatics.irida.repositories.user.UserRepository;
 import ca.corefacility.bioinformatics.irida.ria.web.admin.dto.statistics.GenericStatModel;
 import ca.corefacility.bioinformatics.irida.service.impl.CRUDServiceImpl;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
+
+import com.google.common.base.Strings;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
@@ -469,8 +472,12 @@ public class SampleServiceImpl extends CRUDServiceImpl<Long, Sample> implements 
 			Direction order, String... sortProperties) {
 		sortProperties = verifySortProperties(sortProperties);
 
-		return psjRepository.findAll(ProjectSampleJoinSpecification.searchSampleWithNameInProject(name, project),
-				PageRequest.of(page, size, order, sortProperties));
+		ProjectSampleJoinSpecification psjSampleWithNameInProject = new ProjectSampleJoinSpecification();
+
+		psjSampleWithNameInProject.add(new SearchCriteria("project", project, SearchOperation.EQUAL));
+		psjSampleWithNameInProject.add(new SearchCriteria("sample.sampleName", name, SearchOperation.MATCH));
+
+		return psjRepository.findAll(psjSampleWithNameInProject, PageRequest.of(page, size, order, sortProperties));
 	}
 
 	/**
@@ -555,8 +562,49 @@ public class SampleServiceImpl extends CRUDServiceImpl<Long, Sample> implements 
 	public Page<ProjectSampleJoin> getFilteredSamplesForProjects(List<Project> projects, List<String> sampleNames,
 			String sampleName, String searchTerm, String organism, Date minDate, Date maxDate, int currentPage,
 			int pageSize, Sort sort) {
-		return psjRepository.findAll(ProjectSampleSpecification.getSamples(projects, sampleNames, sampleName,
-				searchTerm, organism, minDate, maxDate), PageRequest.of(currentPage, pageSize, sort));
+
+		ProjectSampleJoinSpecification psjFilteredSamplesForProjects = new ProjectSampleJoinSpecification();
+
+		psjFilteredSamplesForProjects.add(new SearchCriteria("project", projects, SearchOperation.IN));
+
+		// Check to see if the sampleNames are in the samples
+		if (sampleNames.size() > 0) {
+			psjFilteredSamplesForProjects.add(new SearchCriteria("sample.sampleName", sampleNames, SearchOperation.IN));
+		}
+
+		// Check to see if there is a specific sample name
+		if (!Strings.isNullOrEmpty(sampleName)) {
+			psjFilteredSamplesForProjects
+					.add(new SearchCriteria("sample.sampleName", sampleName, SearchOperation.MATCH));
+		}
+
+		// Check for the table search.
+		// This can be expanded in future to search any attribute on the sample (e.g. description)
+		// Underscores within the search term are escaped as the underscores were being treated the same
+		// as hyphens.
+		if (!Strings.isNullOrEmpty(searchTerm)) {
+			psjFilteredSamplesForProjects.add(
+					new SearchCriteria("sample.sampleName", searchTerm.replace("_", "\\_"), SearchOperation.MATCH));
+		}
+
+		// Check for organism
+		if (!Strings.isNullOrEmpty(organism)) {
+			psjFilteredSamplesForProjects.add(new SearchCriteria("sample.organism", organism, SearchOperation.MATCH));
+		}
+
+		// Check if there is a minimum search date
+		if (minDate != null) {
+			psjFilteredSamplesForProjects
+					.add(new SearchCriteria("sample.modifiedDate", minDate, SearchOperation.GREATER_THAN_EQUAL));
+		}
+
+		// Check if there is a maximum search date
+		if (maxDate != null) {
+			psjFilteredSamplesForProjects
+					.add(new SearchCriteria("sample.modifiedDate", maxDate, SearchOperation.LESS_THAN_EQUAL));
+		}
+
+		return psjRepository.findAll(psjFilteredSamplesForProjects, PageRequest.of(currentPage, pageSize, sort));
 	}
 
 	/**
